move rust analyzer integration tests to dir
