+++ tests/rust_analyzer/sample_rust_project/.gitignore
@@ -1,0 +1,1 @@
+target
