+++ tests/rust_analyzer/sample_rust_project/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("Hello, world!");
+}
